Some clean up of unnecessary changes #413
