+++ inventory_management_system_api/main.py
@@ -17,7 +17,7 @@ from inventory_management_system_api.routers.v1 import (
     catalogue_item,
     item,
     manufacturer,
-    settings,
+    setting,
     system,
     unit,
     usage_status,
@@ -93,7 +93,7 @@ app.include_router(manufacturer.router, dependencies=router_dependencies)
 app.include_router(system.router, dependencies=router_dependencies)
 app.include_router(unit.router, dependencies=router_dependencies)
 app.include_router(usage_status.router, dependencies=router_dependencies)
-app.include_router(settings.router, dependencies=router_dependencies)
+app.include_router(setting.router, dependencies=router_dependencies)
 
 
 @app.get("/")
