+++ inventory_management_system_api/routers/v1/setting.py
@@ -1,21 +1,20 @@
 """
-Module for providing an API router which defines routes for managing settings using the `SettingsService`
-service.
+Module for providing an API router which defines routes for managing settings using the `SettingService` service.
 """
 
 import logging
-from typing import Annotated, Optional
+from typing import Annotated
 
-from fastapi import APIRouter, Depends, HTTPException, Path, Query, status
+from fastapi import APIRouter, Depends
 
-from inventory_management_system_api.schemas.settings import SparesDefinitionPutSchema, SparesDefinitionSchema
-from inventory_management_system_api.services.settings import SettingsService
+from inventory_management_system_api.schemas.setting import SparesDefinitionPutSchema, SparesDefinitionSchema
+from inventory_management_system_api.services.setting import SettingService
 
 logger = logging.getLogger()
 
 router = APIRouter(prefix="/v1/settings", tags=["settings"])
 
-SettingsServiceDep = Annotated[SettingsService, Depends(SettingsService)]
+SettingServiceDep = Annotated[SettingService, Depends(SettingService)]
 
 
 @router.put(
@@ -24,7 +23,7 @@ SettingsServiceDep = Annotated[SettingsService, Depends(SettingsService)]
     response_description="Spares definition updated successfully",
 )
 def update_spares_definition(
-    spares_definition: SparesDefinitionPutSchema, settings_service: SettingsServiceDep
+    spares_definition: SparesDefinitionPutSchema, setting_service: SettingServiceDep
 ) -> SparesDefinitionSchema:
     # pylint: disable=missing-function-docstring
     logger.info("Updating spares definition")
@@ -32,5 +31,5 @@ def update_spares_definition(
 
     # TODO: Supposed to have 201 for created and 200 for updated
     # TODO: Appropriate excepts/error logging
-    updated_spares_definition = settings_service.update_spares_definition(spares_definition)
+    updated_spares_definition = setting_service.update_spares_definition(spares_definition)
     return SparesDefinitionSchema(**updated_spares_definition.model_dump())
